refactor(mappings): Add portfolio mapper export
