+++ .vscode/settings.json
@@ -32,6 +32,7 @@
     "longpaths",
     "microfrontend",
     "optionlabel",
+    "pmmmwh",
     "previewpopup",
     "Reenroll",
     "Reenrollment",
