+++ libs/client/components/src/misc/common-toaster/toast-close-button/toast-close-button.tsx
@@ -23,7 +23,7 @@ export function ToastCloseButton({ onClose, closeIconText }: ToastCloseButtonPro
         fontSize="medium"
         aria-hidden="true"
         role="presentation"
-        style={{ color: 'var(--material-color-blue-grey-800)', border: 'none', boxShadow: 'none' }}
+        style={{ color: 'var(--material-color-blue-grey-50)', border: 'none', boxShadow: 'none' }}
       />
     </button>
   );
