+++ libs/client/components/src/misc/common-toaster/toast-content/toast-content.module.scss
@@ -13,7 +13,7 @@
   display: flex;
   align-items: center;
   gap: 5px;
-  color: var(--material-color-blue-grey-800);
+  color: var(--material-color-blue-grey-50);
 
   @media screen and (max-width: 768px) {
     font-size: 0.85rem;
