+++ libs/server/features/portfolios/src/commands/lots/index.ts
@@ -1,0 +1,3 @@
+export * from './create-lot.handler';
+export * from './delete-lot.handler';
+export * from './patch-lot.handler';
