+++ libs/server/features/portfolios/src/commands/portfolios/index.ts
@@ -1,0 +1,3 @@
+export * from './create-portfolio.handler';
+export * from './patch-portfolio.handler';
+export * from './delete-portfolio.handler';
