--- libs/server/features/portfolios/src/controllers/one.spec.ts
@@ -1,7 +1,0 @@
-// this test is for libraries that don;t have any tests yet
-
-describe('One', () => {
-  it('should render successfully', () => {
-    expect(true).toBeTruthy();
-  });
-});
