+++ libs/server/features/portfolios/src/mappings/index.ts
@@ -1,0 +1,1 @@
+export * from './portfolio.mapper';
