+++ libs/server/features/portfolios/src/queries/portfolios/index.ts
@@ -1,0 +1,2 @@
+export * from './get-portfolio.handler';
+export * from './get-portfolios.handler';
