+++ libs/server/middleware/src/index.ts
@@ -1,2 +1,3 @@
 export * from './lib/server-middleware';
 export * from './validators';
+export * from './mediator/mediator';
